+++ com.oracle.truffle.r.test/src/com/oracle/truffle/r/test/builtins/TestBuiltin_strsplit.java
@@ -14,7 +14,7 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  *
  * Copyright (c) 2012-2014, Purdue University
- * Copyright (c) 2013, 2018, Oracle and/or its affiliates
+ * Copyright (c) 2013, 2019, Oracle and/or its affiliates
  *
  * All rights reserved.
  */
@@ -157,5 +157,7 @@ public class TestBuiltin_strsplit extends TestBase {
         assertEval("strsplit('a[1][1]=x11&a[1][2]=x12', ']')");
         assertEval("strsplit('a[1][1]=x11&a[1][2]=x12', '[[]')");
         assertEval("strsplit('a[1][1]=x11&a[1][2]=x12', '[][]')");
+
+        assertEval("strsplit('/some/path/to/somewhere' , '^(?=/)(?!//)|(?<!^)(?<!^/)/', perl = TRUE)");
     }
 }